Sound Task Updates
This update adds 2 new audio sources for the cars mgus and the audio settings for each can be changed. You are able to change the pitch and volume gains as well as the base pitch and volume
