--- Assets/Scenes/SoundTesting.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: ff89cd91da724a449bcddedd82d9d583
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
