--- Assets/Scripts/Under Development/UpdateTextValues.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b600ed756054bd44ab446064be4c41bd
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
